co-locate forgot password form action
