+++ app/_actions/auth-form-actions.ts
@@ -27,25 +27,3 @@ export const signInAction = async (formData: FormData) => {
 
   return redirect(redirectTo)
 }
-
-export const forgotPasswordAction = async (formData: FormData) => {
-  const email = formData.get('email')?.toString()
-  const origin = (await headers()).get('origin')
-  const callbackUrl = formData.get('callbackUrl')?.toString()
-
-  if (!email) {
-    return encodedRedirect('error', '/forgot-password', 'Email is required')
-  }
-
-  const { error } = await tryCatch(authActions.forgotPassword({ email, origin }))
-
-  if (error) {
-    return encodedRedirect('error', '/forgot-password', 'Could not reset password')
-  }
-
-  if (callbackUrl) {
-    return redirect(callbackUrl)
-  }
-
-  return encodedRedirect('success', '/forgot-password', 'Check your email for a link to reset your password.')
-}
